getfiletype.c: Use short for lengths instead of int
The name is 18 bytes, so a 2-byte length rounds off nicely to 20 bytes
without padding.

+++ windows/getfiletype.c
@@ -14,8 +14,8 @@
 #define HASH(x) (int)((x * 0x20010000u)>>29)
 
 static struct {
-    char name[18];
-    int length;
+    char  name[18];
+    short length;
 } types[] = {
     #define E(s) {s "\n", sizeof(s)}
     [HASH(FILE_TYPE_UNKNOWN)] = E("FILE_TYPE_UNKNOWN"),
